Respond to PR feedback/split api/app tests
